refactor!: fix a typo, PlaneDescripter to PlaneDescriptor
