+++ src/foundation/core/GlobalDataRepository.ts
@@ -29,7 +29,7 @@ import Vector3 from '../math/Vector3';
 import MutableMatrix44 from '../math/MutableMatrix44';
 import {WellKnownComponentTIDs} from '../components/WellKnownComponentTIDs';
 import {BoneDataType} from '../definitions/BoneDataType';
-import { ProcessApproachEnum } from '../..';
+import {ProcessApproachEnum} from '../../foundation/definitions/ProcessApproach';
 
 type GlobalPropertyStruct = {
   shaderSemanticsInfo: ShaderSemanticsInfo;
